+++ summon-core/src/commonMain/kotlin/code/yousef/summon/modifier/ModifierExtensions.kt
@@ -52,3 +52,5 @@ fun Modifier.size(value: Number): Modifier =
 fun Modifier.borderRadius(value: Number): Modifier = style("border-radius", "${value}px")
 
 fun Modifier.fontSize(value: Number): Modifier = style("font-size", "${value}px")
+
+fun Modifier.fontSize(value: String): Modifier = style("font-size", value)